Refactor app.js to use routers for assignments and accounts, moving related routes to separate files for better organization and maintainability.
